Create common serial bridge location
