+++ src/apps/common/serial_bridge/serial_bridge.h
@@ -1,0 +1,7 @@
+#ifndef __SERIAL_BRIDGE_H__
+#define __SERIAL_BRIDGE_H__
+
+void serial_bridge_init(void);
+void serial_bridge_handle(void);
+
+#endif
